Test File for Criticality Using Double Arrays
Here is a working file for determining which intersection has the highest priority in each state. The highest priority should reflect the intersection to fix first. 
The 2d array is made of 8 rows and 3 columns i.e. [3][8].
There were some values that were inputted that was the criticality values calculated in MATLAB. 
Below is the table that shows the criticality values for each equipment in each state. This functions returns the index for the row for which equipment has the highest criticality. 

The challenge is to implement this into priority. Higher criticality= first priority. 
0 111    xxx
1 110    xx1
2 101    x1x
3 100    x .008 .008
4 011    1xx
5 010    .005 x .0025
6 001    .0025 .005 x
7 000   not sure yet